Add global TLS cipher suites configurability (#3292)
Ciphers are validated against the allowed ciphers for Envoy. Contour
will exit on startup if any invalid ciphers are present in the config
file. If no ciphers are provided, Contour will use the defaults that
exist now.

Also allows ECDHE-ECDSA-AES(128|256)-GCM-SHA256 ciphers to be passed standalone to help with FIPS support

Fixes #2880

Signed-off-by: Sunjay Bhatia <sunjayb@vmware.com>