internal/envoy/v3: Strip possible port from :authority header
This ensures we don't have to have routes match on a wildcard suffix.
Makes internal Contour logic simpler as we don't have to think about
a possible port in route matches any longer and would make adding
support for wildcard domains a little easier.

Downsides:
- More Lua
- Strips port from header that some application may care about?

Could also apply this change *only* for wildcard routes but it seemed
good to be consistent.

Signed-off-by: Sunjay Bhatia <sunjayb@vmware.com>