_integration: re-enable SNI test (#3435)
The SNI integration test was disabled in #3163
due to unexplained sporadic failures. The failures
have not been reproducible in recent testing. It's
possible that a change made to Contour since #3163
has eliminated the source of the test flake, but
also possible that the test is still flaky and just
did not reproduce during testing.

Closes #2743 (maybe).

Signed-off-by: Steve Kriss <krisss@vmware.com>