makefile: Remove --new from make lint-golint task
When running make checkall locally on my machine, I wouldn't get any errors, but running in CI I would.
If I take out the `--new` flag to the golangci-lint binary I get the matching errors found in CI runs.

My typical workflow is to run a `make checkall` to find any issues before waiting for CI to run, but this
wasn't working.

This removes the `--new` from the local makefile task as well as from the Github Actions job.

Signed-off-by: Steve Sloka <slokas@vmware.com>