cmd/contour: properly configure request/response headers policy
Fixes two issues with configuring request/response headers policy.
The first issue was that the user's configured request headers to
set were inadvertently also being used as the response headers to
set. The second issue was that some harmless empty strings were
being added to the lists of request/response headers to remove.

Signed-off-by: Steve Kriss <krisss@vmware.com>