Refactor Makefile to clarify and make more composable
This commit splits up many of the make targets into sub-steps,
that we can reuse, instead of having the same script called in
multiple places.

I've also removed the old local-dev shortcuts, I think we can do
better.

The intent here is to make it so that we can use the pieces
more easily, eventually aiming at making local dev and bug
reproduction quicker and easier.

Signed-off-by: Nick Young <ynick@vmware.com>