test/e2e: Start of incluster test suite
This test suite should contain tests for things that we have to test
with Contour running in an actual k8s cluster.

This initial commit adds some tests that make sure our Contour
ClusterRole is set up with RBAC for projectcontour.io and Ingress
resources.

Next steps in this track of work are to add more coverage to ensure RBAC
is correct for resources in other API groups and to test other
codepaths/default configuration we rely on in our default deployment
etc.

Signed-off-by: Sunjay Bhatia <sunjayb@vmware.com>