WIP: internal/dag: routes in tcpproxy should result in error
HTTPProxy spec says "If TCPProxy is present, Routes is ignored" but this has
not been true.  Routes are processed and HTTPS upgrade is added for given
routes in HTTP listener (and only HTTP listener).  In this case request for
such route will get "301 Moved Permanently" response.

If routes field does not exist, then HTTP listener may be left unconfigured
unless it is configured due to some other HTTPProxy.  User will get either
connection refused or "404 Not Found" if connecting to insecure port.

The current behavior also has meant that user has needed to define routes
(against documented behavior) if they need HTTPS upgrade for their tcpproxy.

This change is the first phase aiming to conform with the document.

This change will:

* Raise error condition if routes are given (but continue processing them)
* Ensure that there will be HTTP listener with virtualhost matching the
  tcpproxy and catch-all route "/" that will trigger the HTTPS upgrade.

Signed-off-by: Tero Saarni <tero.saarni@est.tech>