hack: Change ginkgo install var to use CI
The install-kubernetes-toolchain.sh installs a few dependencies in order to
build, test, and lint Contour code. The env var to install gingkgo was using
OSType=Linux which breaks on a local linux machine.

This changes that variable to use "CI" which is setup by Github Actions to let the
script know if it's running in Github Actions or not.

Signed-off-by: Steve Sloka <slokas@vmware.com>