internal/envoy: fix TCPProxy weights bug
Fixes a bug where when a TCPProxy had more than one
service defined, any service with a zero weight was
updated to have a weight of 1. The correct behavior
is that if all service have no weights, then they
should be evenly weighted, but if some have non-zero
weights and others have zero weights, they stay as-is.

Signed-off-by: Steve Kriss <krisss@vmware.com>