internal/dag: don't walk DAG to look for service
Removes code that walks the DAG looking for
a matching existing Service when adding a
new Service. This code is very inefficient
when the DAG is large and creates a
performance bottleneck. Some extra memory
may end up being used if duplicate Services
are created, but it's unlikely to be a
material issue in practice.

Closes #4058.

Signed-off-by: Steve Kriss <krisss@vmware.com>