Some test additions and fixing script duplication
Unfortunately I have not been able to get a global Lua script working
with both per-Cluster and per-Route policies. Once you configure a
per-Cluster Lua script it supercedes the per-Route configuration and the
per-Route one is not run. Instead, the Lua is always on each Cluster so
it is at the lowest level, per-Route and per-Cluster policies are
aggregated and rewrites happen in one place.

Signed-off-by: Sunjay Bhatia <sunjayb@vmware.com>