Refactor install-contour-working and kind-load scripts
The former is no longer used in CI so simplify to just build local code
and load. The latter is used in CI but should be usable for local
development iteration as well. Ensures make e2e works without setting
env vars, and will build local copy of code. Also can be used with env
vars for CI to use a pre-built image.

Signed-off-by: Sunjay Bhatia <sunjayb@vmware.com>