Bump Envoy to v1.20.1 (#4075)
See release notes:
https://www.envoyproxy.io/docs/envoy/v1.20.1/version_history/current

Also bumps go-control-plane to latest

Deprecation of existing headermatch usage will need to be handled in the release after this is
merged. The "string match" header matching method is preferred but was
added in 1.20.1 and it was not present in 1.19.x, we will need to
release a version with Envoy 1.20.0 compatibility and the subsequent
release will need to move to the "string match" method (rather than the
individual regex/exact/etc. methods.

Signed-off-by: Sunjay Bhatia <sunjayb@vmware.com>