hack/release: Remove ttlSecondsAfterFinished from certgen job (#4200)
The certgen job is versioned in releases, so deleting it should not be
necessary for most consumers. Deleting the job can cause problems when
applying the quickstart config via continous deployment (e.g. Argo CD)
as the job disappears before an external observer can acknowledge it
completed successfully, or even confirm that it ever existed.

Signed-off-by: Luke Rewega <lrewega@buf.build>