Gateway API: drop 'core' as allowed core API group name
Previously Contour accepted either the empty string
or "core" as the group name for the core API group,
i.e. the one containing Services, Secrets, etc. However,
the "core" string is not part of the Gateway API spec
and adds complexity to all logic that interacts with
groups. This PR drops support for it, and expects
that any references to the core API group will have
an empty group string.

Signed-off-by: Steve Kriss <krisss@vmware.com>