internal/k8s: StatusUpdateWriter only writes to update channel when leader
Otherwise, when not leader, update buffered channel can be filled up,
since StatusUpdateHandler does not run. This deadlocks the EventHandler.

Shows up in e2e tests in intermittent in-cluster smoke-test failures.
More prevalent with more instances so modified smoke test to test more
apps. The failure occurs when the contour instance Envoy connects to is
not the leader.

Signed-off-by: Sunjay Bhatia <sunjayb@vmware.com>