internal/contour: Removed duplicated function for secret validation in dag layer
Removed validSecret function from accessors.go in the dag package.
This function checked if a secret is a valid TLS secret and returned an error.
Substituted with the validTLSSecret in secret.go in the dag package,
function that wraps the isValidSecret function by adding some logic on it.
This way the duplication of code is avoided and the feature of the deleted function is maintained.

Fixes #3503

Signed-off-by: alessandroargentieri <alexmawashi87@gmail.com>