demote Contour CRD to an internal type that serves as a model
Note, this also removes logic that previously checked
for the existence of other Contour CRs in a given ns,
or in the cluster, for the purposes of avoiding issues
with shared resources. We need to implement something
similar for Gateway, but the basic assumption right
now is that you can only have one Contour per ns so
we should avoid instantiating more than one per ns
to begin with.

Signed-off-by: Steve Kriss <krisss@vmware.com>