validate Gateway HTTP listener ports/hostnames
Adds validation to both the Gateway provisioner
and the Gateway API DAG processor to check HTTP
protocol listeners to ensure a single port, and
no hostname conflicts. The provisioner uses this
to configure the correct port on the Envoy service,
and the DAG processor uses it to set listener
conditions and process only valid listeners.

Signed-off-by: Steve Kriss <krisss@vmware.com>