internal/envoy: Fix disableMergeSlashes config option propagation
In 335e816 we made Envoy's
merge_slashes option configurable via the ContourConfiguration CRD and
the main configuration file, but the value from the configuration isn't
actually propagated to listeners. I did not notice this in our original
testing pass as that was performed before the polarity of the config
option was inverted.

As a fix, ensure we correctly pass the disableMergeSlashes config option
value to Envoy listener configs.

Signed-off-by: Máté Szabó <mszabo@fandom.com>