CORSPolicy AllowOrigin can be configured as a regex
This will allow more flexible CORS Allowed Origins with a regex pattern,
rather than having to allow all (w/ "*") or enumerating all exact
matches.

We attempt to parse each allow origin match as an Origin header (using
go url parsing) to check if we get a scheme://host[:port]. If the string
doesnt parse as a URL or the format is invalid for an Origin header
exact match, then we attempt to parse it as a regex.

Fixes: https://github.com/projectcontour/contour/issues/4666

Signed-off-by: Sunjay Bhatia <sunjayb@vmware.com>