remove shutdown-manager liveness probe
The probe can currently cause problems when it fails
by causing the shutdown-manager container to be restarted
by itself, which then results in the envoy container
getting stuck in a "DRAINING" state indefinitely.

Not having the probe is less bad overall because envoy pods
are less likely to get stuck in "DRAINING", and the
worst case without it is that shutdown-manager is truly
unresponsive during a pod termination, in which case
the envoy container will simply terminate without first
draining active connections.

Updates #4851.

Signed-off-by: Steve Kriss <krisss@vmware.com>