Fix the upgrade e2e test flakiness
The upgrade e2e test can fail in some situations because the envoy
waiter is incorrectly returning true before the StatefulSet even manages
to finish the rollout since it only checks the StatefulSet state on
properties that are inconsistently populated due to timing differences.

This fix addressed that by using a similar condition validation method
as the one already used for the Contour Deployment rollout.

Signed-off-by: Aurel Canciu <aurelcanciu@gmail.com>