test: use helper to generate ObjectMeta
Test objects were created with uninitialized ResourceVersion and Generation
since the values did not matter when complete object content was compared to
detect updates.

This change is preparation for #5064 and introduces a helper function to
generate unique ObjectMeta. Tests cases that depend on object update are
changed to use the helper instead of hardcoded initialization.

Signed-off-by: Tero Saarni <tero.saarni@est.tech>