internal/sorter: Routes with method matching get precedence (#5434)
This is to support Gateway API conformance, release 0.7.1 dictates that
HTTP method comes before length of header matches in the precendence
order for choosing routes.

Since we currently implement method matching in the DAG by adding a
:method header match, this solution finds any :method headers being
matched on and first determines precendence of two routes being compared
based on that. If method matches are equal (both present or not
present), then we go ahead with comparing the remaining header matches.

Since HTTPProxy provides this ability as well, this will affect priority
of routes from HTTPProxies, not just Gateway API HTTPRoute.

Alternately we could add a method field directly to the dag.Route struct
and implement method matching that way. We may want to in that case also
sanitize the header matches to make sure we properly collect matches on
the :method header, or disallow using them in HTTPRoute, etc.

Signed-off-by: Sunjay Bhatia <sunjayb@vmware.com>