internal/dag: GetServiceClusters no longer returns ExternalName clusters (#5472)
This method is used by the EndpointsTranslator on dag change to
recalculate the relevant endpoints to send to Envoy. Since we do not use
EDS to configure endpoints for Clusters that come from ExternalName
Services, we no longer collect ServiceClusters associated with these
Clusters. This is required for ADS, as go-control-plane will not respond
to requests if the snapshot contains resources that were not requested.

Signed-off-by: Sunjay Bhatia <sunjayb@vmware.com>