Adds support for treating missing headers as empty
`TreatMissingHeadersAsEmpty` specifies if the header match rule specified header does not exist,
this header value will be treated as empty. Defaults to false.
Unlike the underlying Envoy implementation this is **only** supported
for negative matches (e.g. NotContains, NotExact).

The reason that I implemented only for negative matches is that is
substantially simpler to reason about. I am open to implementing it
for all cases if you think it is going to be useful.

Signed-off-by: Sotiris Nanopoulos <sotiris.nanopoulos@reddit.com>