Stops route sorting on `HTTPProxy` CRD
Contour generates the Envoy routes for each virtual host in the same
order as they are in the `HTTPProxy`. This benefits users from having no
suprises between the Envoy actual routing table and what they described in the
CRD.

* Change the underlying data structure in `dag.go` `VirtualHost` to be a slice
such that we can maintain the order.
* Change the sorting to be conditional on the caller wanting it. This is used to implement sorting
for virtual hosts which are "partially" controlled by an Ingress so we can have ingress conformance.
* Change the implementation of `expandPrefixMatches` to maintain the order of the original list.

mechanical to switch it to do it.

I don't have a strong opinion. I would prefer if we didnt but I am not familiar with any conformance requirements.

I believe this would be a lot of work given the tests but I am not sure what the spec requires.

Signed-off-by: Sotiris Nanopoulos <sotiris.nanopoulos@reddit.com>