Remove testing around V().Info()
logrusr has changed behavior since v3.0.0, it now tries to mimic logrus
log levels with the V() level, see:
https://github.com/bombsimon/logrusr/commit/9f3fd50a0eadf8ee1cecbe31e6d6b232ac40a283

In practice client-go checks if a certain verbosity level is enabled and
initializes a different logger based on that and then uses Info(f) logs,
rather than the V().Info() construction.

This commit removes the testing of log lines written with V() guarding
them and rather just tests the expected verbosity is enabled or not.

Signed-off-by: Sunjay Bhatia <sunjayb@vmware.com>