internal/envoy: use consistent HTTP filter names
Use the same names for HTTP filters when they
are added to an HTTP connection manager and
when they are configured with per-filter config
on virtual hosts or routes. Extracts a set of
constants for filter names to ensure they remain
in sync.

This addresses a change in Envoy behavior in Envoy
1.29. See https://github.com/envoyproxy/envoy/issues/29461
for more detail.

Signed-off-by: Steve Kriss <stephen.kriss@gmail.com>