docs: Update README.md to be more helpful (#6585)
The docs/README.md made no sense. Anyone reading it in GitHub clearly wants to
contribute to the documentation, that's why they're in the source code of
Contour, why else would they have found their way to the source repository? So,
it should point to where the documentation lives in the git repository, not to
the website where it's served.

Signed-off-by: James Roper <james@jazzy.id.au>
Co-authored-by: Steve Kriss <stephen.kriss@gmail.com>