Removed +required tag from ExtensionServiceReference.Name
Starting with version v0.16.0 of controller-tool, the +required tag began being processed, leading to the inclusion of new "required" entries in CRDs.

The ExtensionServiceReference is used in the AuthorizationServer as an optional field, but it is embedded as a struct rather than a pointer.
This causes it to be included in requests even when not explicitly set.

This resulted in the following error:

HTTPProxy.projectcontour.io "external-auth" is invalid:
[spec.virtualhost.authorization.extensionRef.name: Required value, <nil>:
Invalid value: "null": some validation rules were not checked because the
object was invalid; correct the existing errors to complete validation]

Ideally, HTTPProxy.Spec.VirtualHost.Authorization.ExtensionServiceRef should have been a pointer to make it truly optional.
However, changing it to a pointer now would break backward compatibility.
An alternative solution is to make ExtensionServiceReference.Name optional, which should be acceptable since the +required tag was never enforced by the CRD in the first place.

Signed-off-by: Tero Saarni <tero.saarni@est.tech>