Change conflicting FQDN logic in HTTPProxy
Sort HTTPProxies when checking for FQDN conflicts so that the
oldest is still set as Valid so multiple tenants cannot affect
each other, e.g. resources across namespaces should not be able
to invalidate the original "owner" of the FQDN.

Useful for cases where Gatekeeper policies are not feasible to
apply.

Signed-off-by: Sunjay Bhatia <sunjay.bhatia@broadcom.com>